+++ client/jetbrains/CHANGELOG.md
@@ -47,6 +47,7 @@
 - Fixed line breaks in the chat ui [#53543](https://github.com/sourcegraph/sourcegraph/pull/53543)
 - Reset prompt input on message send [#53543](https://github.com/sourcegraph/sourcegraph/pull/53543)
 - Fixed UI of the prompt input [#53548](https://github.com/sourcegraph/sourcegraph/pull/53548)
+- Fixed zero-width spaces popping up in inline autocomplete [#53599](https://github.com/sourcegraph/sourcegraph/pull/53599)
 
 ### Security